backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
supabase==2.3.1
pydantic==1.10.15
python-dotenv==1.0.1
httpx ==0.25.2
Pillow==10.4.0
rembg==2.0.57
pdfplumber==0.11.2
pytesseract==0.3.10
sentence-transformers==3.0.1
numpy==1.26.4
openai-whisper==20240930
requests==2.32.3
orjson==3.10.7
python-multipart==0.0.9
itsdangerous
gotrue==2.4.1
gunicorn